# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: stage
          in: query
          required: false
          schema: { $ref: "#/components/schemas/LeadStage" }
        - name: source
          in: query
          required: false
          schema: { $ref: "#/components/schemas/LeadSource" }
        - name: insuranceType
          in: query
          required: false
          schema: { $ref: "#/components/schemas/InsuranceType" }
        - name: city
          in: query
          required: false
          schema: { type: string }
        - name: assignedTo
          in: query
          required: false
          schema: { type: string }
        - name: dateFrom
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: dateTo
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [created_at, last_contacted_at, next_followup_at]
        - name: sortDir
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
        - name: page
          in: query
          required: false
          schema: { type: integer }
        - name: pageSize
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadList"
    post:
      operationId: createLead
      tags: [leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetail"
    patch:
      operationId: updateLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /leads/{id}/timeline:
    get:
      operationId: getLeadTimeline
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
  /leads/{id}/notes:
    post:
      operationId: addLeadNote
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}/call:
    post:
      operationId: triggerLeadCall
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerCallInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerCallResult"
  /leads/bulk-import:
    post:
      operationId: bulkImportLeads
      tags: [leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"
...
components:
  schemas:
...
    Gender:
      type: string
      enum: [MALE, FEMALE, OTHER]
    InsuranceType:
      type: string
      enum: [LIFE, HEALTH, MOTOR, TERM, ULIP, ENDOWMENT, ACCIDENT, TRAVEL]
    LeadSource:
      type: string
      enum: [META_ADS, WEBSITE_FORM, CSV_UPLOAD, MANUAL, INBOUND_CALL, REFERRAL]
    LeadStage:
      type: string
      enum: [NEW, CONTACTED, INTERESTED, DOCS_PENDING, POLICY_ISSUED, RENEWAL_DUE, LOST, DO_NOT_CALL]
    PolicyStatus:
      type: string
      enum: [ACTIVE, LAPSED, SURRENDERED, MATURED, CLAIMED]
    PremiumFreq:
      type: string
      enum: [MONTHLY, QUARTERLY, HALF_YEARLY, YEARLY, SINGLE]
    CallDir:
      type: string
      enum: [OUTBOUND, INBOUND]
    CallStatus:
      type: string
      enum: [INITIATED, RINGING, IN_PROGRESS, COMPLETED, FAILED, NO_ANSWER, BUSY, CANCELLED]
...
    FollowUpType:
      type: string
      enum: [RENEWAL_REMINDER, MONTHLY_CHECKIN, CALLBACK_REQUESTED, MANUAL, POLICY_ANNIVERSARY]
    FollowUpStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, COMPLETED, SKIPPED, RESCHEDULED]
...
    Lead:
      type: object
      required: [id, org_id, full_name, gender, phone, source, stage, tags, created_at, updated_at]
      properties:
        id: { type: string }
        org_id: { type: string }
        full_name: { type: string }
        gender: { $ref: "#/components/schemas/Gender" }
        dob: { type: ["string", "null"], format: date-time }
        age: { type: ["integer", "null"] }
        phone: { type: string }
        phone_alt: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address_line1: { type: ["string", "null"] }
        address_line2: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        pincode: { type: ["string", "null"] }
        pan_number: { type: ["string", "null"] }
        aadhaar_number: { type: ["string", "null"] }
        occupation: { type: ["string", "null"] }
        annual_income: { type: ["integer", "null"] }
        employer_name: { type: ["string", "null"] }
        insurance_type: { oneOf: [ { $ref: "#/components/schemas/InsuranceType" }, { type: "null" } ] }
        sum_assured_interest: { type: ["integer", "null"] }
        premium_budget: { type: ["integer", "null"] }
        source: { $ref: "#/components/schemas/LeadSource" }
        source_campaign_id: { type: ["string", "null"] }
        source_form_id: { type: ["string", "null"] }
        stage: { $ref: "#/components/schemas/LeadStage" }
        assigned_to: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        last_contacted_at: { type: ["string", "null"], format: date-time }
        next_followup_at: { type: ["string", "null"], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    LeadDetail:
      allOf:
        - $ref: "#/components/schemas/Lead"
        - type: object
          properties:
            policies:
              type: array
              items: { $ref: "#/components/schemas/Policy" }
            call_logs:
              type: array
              items: { $ref: "#/components/schemas/CallLog" }
            follow_ups:
              type: array
              items: { $ref: "#/components/schemas/FollowUp" }
    LeadInput:
      type: object
      required: [full_name, phone]
      properties:
        full_name: { type: string, minLength: 1 }
        gender: { $ref: "#/components/schemas/Gender" }
        dob: { type: string, format: date-time }
        age: { type: integer }
        phone: { type: string }
        phone_alt: { type: string }
        email: { type: string }
        address_line1: { type: string }
        address_line2: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        pan_number: { type: string }
        aadhaar_number: { type: string }
        occupation: { type: string }
        annual_income: { type: integer }
        employer_name: { type: string }
        insurance_type: { $ref: "#/components/schemas/InsuranceType" }
        sum_assured_interest: { type: integer }
        premium_budget: { type: integer }
        source: { $ref: "#/components/schemas/LeadSource" }
        stage: { $ref: "#/components/schemas/LeadStage" }
        assigned_to: { type: string }
        notes: { type: string }
        tags:
          type: array
          items: { type: string }
        next_followup_at: { type: string, format: date-time }
    LeadUpdate:
      type: object
      properties:
        full_name: { type: string }
        gender: { $ref: "#/components/schemas/Gender" }
        dob: { type: ["string", "null"], format: date-time }
        age: { type: ["integer", "null"] }
        phone: { type: string }
        phone_alt: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address_line1: { type: ["string", "null"] }
        address_line2: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        pincode: { type: ["string", "null"] }
        pan_number: { type: ["string", "null"] }
        aadhaar_number: { type: ["string", "null"] }
        occupation: { type: ["string", "null"] }
        annual_income: { type: ["integer", "null"] }
        employer_name: { type: ["string", "null"] }
        insurance_type: { oneOf: [ { $ref: "#/components/schemas/InsuranceType" }, { type: "null" } ] }
        sum_assured_interest: { type: ["integer", "null"] }
        premium_budget: { type: ["integer", "null"] }
        stage: { $ref: "#/components/schemas/LeadStage" }
        assigned_to: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        next_followup_at: { type: ["string", "null"], format: date-time }
    LeadList:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Lead" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    NoteInput:
      type: object
      required: [note]
      properties:
        note: { type: string, minLength: 1 }
    TriggerCallInput:
      type: object
      required: [agent_id]
      properties:
        agent_id: { type: string }
    TriggerCallResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        call_log_id: { type: ["string", "null"] }
        execution_id: { type: ["string", "null"] }
        error: { type: ["string", "null"] }
    BulkImportInput:
      type: object
      required: [rows]
      properties:
        trigger_calls: { type: boolean }
        agent_id: { type: string }
        rows:
          type: array
          items: { $ref: "#/components/schemas/BulkImportRow" }
    BulkImportRow:
      type: object
      properties:
        full_name: { type: string }
        phone: { type: string }
        gender: { type: string }
        dob: { type: string }
        email: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        insurance_type: { type: string }
        notes: { type: string }
    BulkImportResult:
      type: object
      required: [imported, skipped_duplicates, skipped_invalid, errors]
      properties:
        imported: { type: integer }
        skipped_duplicates: { type: integer }
        skipped_invalid: { type: integer }
        errors:
          type: array
          items: { type: string }
    Policy:
      type: object
      required: [id, org_id, lead_id, policy_type, premium_frequency, status, created_at, updated_at]
      properties:
        id: { type: string }
        org_id: { type: string }
        lead_id: { type: string }
        lead_name: { type: ["string", "null"] }
        policy_number: { type: ["string", "null"] }
        insurer_name: { type: ["string", "null"] }
        policy_type: { $ref: "#/components/schemas/InsuranceType" }
        sum_assured: { type: ["integer", "null"] }
        annual_premium: { type: ["integer", "null"] }
        premium_frequency: { $ref: "#/components/schemas/PremiumFreq" }
        start_date: { type: ["string", "null"], format: date-time }
        end_date: { type: ["string", "null"], format: date-time }
        renewal_date: { type: ["string", "null"], format: date-time }
        nominee_name: { type: ["string", "null"] }
        nominee_relation: { type: ["string", "null"] }
        nominee_dob: { type: ["string", "null"], format: date-time }
        status: { $ref: "#/components/schemas/PolicyStatus" }
        documents:
          type: array
          items: { $ref: "#/components/schemas/PolicyDocument" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PolicyDocument:
      type: object
      required: [name, url]
      properties:
        name: { type: string }
        url: { type: string }
...
    CallLog:
      type: object
      required: [id, org_id, bolna_execution_id, bolna_agent_id, direction, phone_number, status, started_at, drop_detected, created_at, updated_at]
      properties:
        id: { type: string }
        org_id: { type: string }
        lead_id: { type: ["string", "null"] }
        lead_name: { type: ["string", "null"] }
        bolna_execution_id: { type: string }
        bolna_agent_id: { type: string }
        agent_name: { type: ["string", "null"] }
        direction: { $ref: "#/components/schemas/CallDir" }
        phone_number: { type: string }
        status: { $ref: "#/components/schemas/CallStatus" }
        started_at: { type: string, format: date-time }
        ended_at: { type: ["string", "null"], format: date-time }
        duration_seconds: { type: ["integer", "null"] }
        transcript: { type: ["string", "null"] }
        summary: { type: ["string", "null"] }
        recording_url: { type: ["string", "null"] }
        disposition_id: { type: ["string", "null"] }
        disposition: { oneOf: [ { $ref: "#/components/schemas/Disposition" }, { type: "null" } ] }
        drop_detected: { type: boolean }
        drop_reason: { type: ["string", "null"] }
        retry_of_call_id: { type: ["string", "null"] }
        retry_call_id: { type: ["string", "null"] }
        call_type: { type: ["string", "null"] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Disposition:
      type: object
      required: [id, org_id, bolna_agent_id, label, color, is_active, sort_order, created_at]
      properties:
        id: { type: string }
        org_id: { type: string }
        bolna_agent_id: { type: string }
        label: { type: string }
        color: { type: string }
        description: { type: ["string", "null"] }
        is_active: { type: boolean }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
...
    FollowUp:
      type: object
      required: [id, org_id, lead_id, type, scheduled_at, status, created_at, updated_at]
      properties:
        id: { type: string }
        org_id: { type: string }
        lead_id: { type: string }
        lead_name: { type: ["string", "null"] }
        policy_id: { type: ["string", "null"] }
        type: { $ref: "#/components/schemas/FollowUpType" }
        scheduled_at: { type: string, format: date-time }
        completed_at: { type: ["string", "null"], format: date-time }
        bolna_agent_id: { type: ["string", "null"] }
        call_log_id: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/FollowUpStatus" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    TimelineEvent:
      type: object
      required: [id, kind, title, timestamp]
      properties:
        id: { type: string }
        kind: { type: string, enum: [call, stage_change, note, source] }
        title: { type: string }
        detail: { type: ["string", "null"] }
        timestamp: { type: string, format: date-time }
...